fixed collision but, now the enemy flees if they collide to each other

+++ DungeonDrive/DungeonDrive/Unit.cs
@@ -8,8 +8,8 @@ namespace DungeonDrive
         public double x;
         public double y;
         public double speed = 0.01;
-        public int hp = 1;
-        public int atk_dmg = 1;
+        public double hp = 1;
+        public double atk_dmg = 1;
 
         public int DrawX { get { return (int)(x * G.size + G.width / 2 - G.hero.x * G.size - G.size / 2); } }
         public int DrawY { get { return (int)(y * G.size + G.height / 2 - G.hero.y * G.size - G.size / 2); } }
